Revert "Feature/params"
