+++ templates/basic/.eslintignore
