+++ templates/basic/.eslintrc.json
