+++ templates/basic/.gitignore
