+++ templates/basic/babel.config.json
