+++ templates/basic/config/webpack.common.js
