+++ templates/basic/config/webpack.dev.js
