+++ templates/basic/config/webpack.prod.js
