+++ templates/basic/package.json
