+++ templates/basic/src/App.tsx
