+++ templates/basic/styles/GlobalStyles.ts
