+++ templates/basic/tsconfig.json
