--- templates/schemas/component/ts/index.tsx
@@ -1,12 +1,0 @@
-import {} from './styles'
-
-interface ComponentProps {
-
-}
-
-export const Component = (props: ComponentProps): JSX.Element  => {
-  return (
-    <>
-    </>
-  )
-}
