--- templates/schemas/component/ts/styles.ts
@@ -1,2 +1,0 @@
-import styled from 'styled-components'
-
