--- templates/schemas/redux-component/ts/actions.ts
@@ -1,12 +1,0 @@
-/*
-  An Example of an action
-
-  import { Action, TYPE } from "./types"
-  
-  export const test = (): Action => ({
-    type: TYPE,
-    payload: {
-      foo: 'bar'
-    }
-  })
-*/
