--- templates/schemas/redux-component/ts/reducer.ts
@@ -1,11 +1,0 @@
-import { Action, State } from './types';
-
-const initialState: State = {
-}
-
-export const Reducer = (state = initialState, action: Action) => {
-  switch(action.type) {
-    default:
-      return state;
-  }
-}
